feat : 온보딩시 입력받은 닉네임 저장 기능 구현 refactor : 헥사고날 구조에 맞게 port, adapter 패키지 수정
