+++ booquest-api/src/main/java/com/booquest/booquest_api/application/port/in/user/UpdateUserProfileUseCase.java
@@ -1,0 +1,5 @@
+package com.booquest.booquest_api.application.port.in.user;
+
+public interface UpdateUserProfileUseCase {
+    void updateNickname(long userId, String nickname);
+}
